Tie off loose ends
This whole idea is rubbish.  There's no way to communicate the setting information
via the H-M spacegroup symbol in this way, except in a few circumstances.
See ITA (2005) section 4.3.

This simplifies matters a lot.  There's no need for any extra parsing, because
the "_uaX" system is sufficient, and is the best that can be done without
inventing a whole new system for point group symbols.