Iteration over reflection list should include other reflections with same asymmetric indices
This fixes a >2 year old bug in the reflection list, and a bug in tests/list_check.c which
allowed it to go undetected for so long.  The main situation where it would have caused
problems is when asymmetric_indices() was called on a reflection list containing symmetrically
equivalent reflections.  The only core CrystFEL program which does this is partialator.
compare_hkl uses asymmetric_indices(), but the reflections are already guaranteed to be
asymmetric by that point (because check_list_symmetry() has been called).