partialator: New convergence criterion for scaling
Instead of stopping iteration when the absolute value of any scaling factor changes by more
than a certain (small) amount, calculate the ratios of the scaling factors to their
previous values, and stop when no scaling factor changes by more than 1% compared to the
mean ratio.

This method is robust against "drifting" of the scale factors when the partiality
estimates are poor.