"Retry" mechanism for indexing
This increases the indexing rate a bit in situations where there are
lots of weak peaks which, although they may be real, don't help
indexing.  The weakest 10% of peaks get cut out and the indexing re-run.

This also allows multiple hits to be indexed, using the "inelegant peak
subtraction method", by retrying indexing in the same way after deleting
peaks which are accounted for by the lattice just found.