Perform prediction refinement straight after indexing
This allows indexing to be attempted again (either a new method or with
"retry") if the prediction refinement fails, increasing overall indexing
rate.

Side-effect: there are some hoops which would need to be jumped through
to store the profile radius before refinement and hence enable
scripts/plot-predict-refine to work.  For now, we'll ignore this as it's
clear that the prediction refinement is working.