Allow indexing system to store its own data, independently of indexing methods
Previously, the indexing system passed all the information on to the
indexing engines and then forgot about it.  That made it difficult to do
things like check the indexing solution after prediction refinement,
because the target unit cell was unavailable.  Now, the indexing system
itself can keep some information.  Of course, that information includes
the private pointers for the indexing engines themselves.

I took the opportunity to streamline things a little bit.  The caller
can now set up the indexing system in one step, without having to
separately parse the names of the indexing methods.  The caller no
longer has to keep track of a separate array of methods, instead just
one structure which contains everything.