Offset peak locations from HDF5 or CXI files by 0.5,0.5
CrystFEL considers all peak locations to be distances from the corner of
the detector panel, in pixel units, consistent with its description of
detector geometry.  In contrast, Cheetah considers the peak locations
to be pixel indices in the data array.  Therefore, a half-pixel offset
is needed when importing the peak lists.

For users who need the old behaviour, this commit adds a new option
indexamajig --no-half-pixel-shift to deactivate this offset.