Fill in photon energy, clen and adu for CBFs
This needed a bit of reorganisation and clarification of who is
repsonsible for loading what.  The low-level file loaders, e.g.
hdf5_read and hdf5_read2 in hdf5-file.c or cbf_read in image.c, are
responsible.  There is a helper function adjust_centering_for_rail in
detector.h which they can use.

It seems like this could be done more cleanly at the imagefile layer.
However, we need to make sure that the "old" hdfile API still works on
its own, even when not accessed via the new imagefile API.