Make all Python scripts compatible with Python 2 and 3
It would be nice to upgrade completely to Python 3, but this may
create problems at SLAC because psana only supports Python 2 (Python 3
is available, but means that users have to jump through hoops to run
a simple CrystFEL script).  None of our scripts do anything complicated,
so they can all be compatible with both so far.

If anyone adds a script which requires a particular version, make sure
to specify the version in the first line, consistent with PEP 394.