Accept either cbf/cbf.h or cbflib/cbf.h
You get cbflib/cbf.h if you install CBFlib from source, and cbf/cbf.h if
using the package manager in Fedora, probably other systems as well.

We definitely don't want to include the cbf or cbflib folder in the
include path, because CBFlib may bring its own HDF5 headers, which we
don't want to touch with a bargepole.