Simplify indexing options
mosflm-cell-nolatt-multi-refine-bad-retry.... it's all getting a bit
too much.  This commit removes all the options from the indexing methods
except for the prior information flags (cell/nocell and latt/nolatt).
All the other options are now given at the top level with new
indexamajig options, e.g. --no-multi and --no-retry.  They are applied
to all the methods, with no option to (say) use retry for one method and
not for another.

This also makes the "peak alignment test" default to OFF, and
multi-lattice indexing ON.  I think that the prediction refinement does
a good enough job of checking the validity of indexing solutions,
allowing us to turn off the alignment check which filters out most
multi-lattice solutions.