Remove write_reflections_to_file() from API
It's there because it used to be used to write the reflections into the
stream.  Therefore there was a need for a function to write a bare list
of reflections to a file, with no headers.  Since then, the stream and
merged reflection file formats have diverged, so there's no use for this
other than to confuse people.