Embed diffraction.cl and hdfsee.ui into executables
Having to install these causes problems, e.g. you can't run from build
directory nor run tests without installing.

For hdfsee.ui, the XML is now embedded in the code just as it is for
cell_explorer.

diffraction.cl is a bit too long and complicated for that.  There are a
number of nasty ways to embed such files.  This one seems to be the
overall least nasty, although it requires data/gen-resources to be run
whenever diffraction.cl is changed.