uncenter_cell(): Convert R to P
Otherwise, the cell comparison doesn't recognise the cells as the same,
breaking indexing with an R reference cell.

Unfortunately, the resulting cell after comparison by compare_r_c_p
still comes out as P, but that doesn't seem like a big loss.  To get it
strictly correct, we'd need some way of tracking through the information
that the cell got "uncentered" from R to P, even though the matrix is an
identity.