Break down hdfile_get_string_value()
There are just too many cases here.  Variable/fixed length, array of
either of those, with/without event structure.  And that's just for real
strings (not numbers that will be converted to strings).

For now, I think the only unhandled case is an array of variable length
strings.  We can fix this when/if one turns up in the wild.