pinkIndexer: Remove detector update
Overall detector offsets can now be done via the geometry file, and this
is labelled as a hack.  It doesn't actually do very much anyway, because the
"updated" detector is thrown away for the next pattern.  The updated detector
is used only for subsequent indexing attempts on the same pattern.