Keep a separate record of unit cell file
There seems to be no way to get a clear answer from the GtkFileChooser
about whether a file is actually set.  Instead, it now keeps a separate
record which is set to NULL when the "clear" button is pressed, and set
to the filename from the GtkFileChooserButton on the "file-set" signal.