partialator: Reject crystals with obviously too large profile radii
The criterion for "too large" is 20% of the 1/d value for the lowest
reflection which is not systematically absent according to the
centering.

A profile radius larger than the 1/d value for a reflection will crash
the xsphere partiality model, and some visualisation shows that this is
a clearly non-physical situation.  The profile radius shouldn't be
anywhere near the inter-Bragg spacing for reasonable data.

However, feedback shows that this is happening quite often in real data,
probably due to bad indexing.