UnitCell: Store all representations once they're calculated
Previously, the "getter" functions would re-calculate the requested
representation every time they were called.  This could mean doing a
matrix inversion in the middle of a tight loop, wasting loads of time.
Now, it stores the calculated values and returns them directly next
time.  Setting the parameters invalidates the values for all
representations other than the one given.

The cost of doing this is that the cell can no longer be "const" in the
getter functions.  This tracked through some other code, but nothing too
severe.