check_hkl: Move "single-list" FoMs into API
Reasons for differences:

1. Resolution shells slightly different
The binning calculation needs to take into account small rounding errors
in the resolution calculation, when not using an explicit resolution
range (--highres).  The old version did this by taking a min/max
resolution range slightly larger than the resolution of the data.  The
new version handles the rounding errors explicitly, so does not need
this.

2. Number of reflections with infinite/invalid I/sigI values halved
The number reported for this count was twice what it should have been,
due to a bug in the old check_hkl.

3. Overall SNR is different
When the above warning message applies, the old version still allowed
the reflections with invalid I/sigI values to contribute to the
denominator of the mean SNR calculation.  The new version does not
include them in the SNR calculation at all.  Note that the reflections
contribute to the other figures of merit unless otherwise stated.

4. Standard deviation of intensity is not calculated
It would've been a lot of work to include this in the new version, and
it's a totally useless number.  If you disagree, please get in touch!