GUI: Back out export reindex code
The reindexing is really going to create a whole world of pain.  Even if
we can make it work correctly, it still might come as a surprise to
people when the exported data is indexed differently (e.g. to data
exported to some other format).  Let's just forget the whole thing, and
refuse to export to formats which don't have a way to represent the data
we have.  Perhaps revisit this later if it becomes a big problem and we
want to put in the effort required.