GUI: Re-work local backend plumbing
A single command line is not expressive enough for at least two cases:
process_hkl, which needs to be run two extra times to generate the splits,
and ambigator, which needs to have its input streams concatenated first.

This also splits out the routine for creating the working folder and
writing the notes file.