indexamajig: Add --camera-length-estimate
Unfortunately, PinkIndexer needs the real camera length for its centre
refinement.  Giving a fake value and scaling the resulting shift does
not work - the indexing rate drops with even a small error.

Ideally, this would work in the same way as --wavelength-estimate, by
using a static value from the geometry file if it's given.  However,
this is rather complicated to implement because of the way all the units
stuff is implemented.  Therefore, this is left as an improvement for the
future.