indexamajig: Generate a unique filename for ZMQ data
The unique filename is needed by the GUI for looking up results in a
stream.  Otherwise, the filename is "(null)" for everything and the
lookup just returns the first chunk in the stream.

The filename is generated based on the unique serial number for each
chunk, and is therefore unique across one run of indexamajig regardless
of the number of worker processes (-j).  This might have to change in future
to accommodate jobs run across multiple nodes, if there is any demand
for looking at results in one big concatenated stream.

This also changes the condition for deciding when to look for a 'real'
file, to take into account that there is always a non-NULL filename.