GUI: When indexing one frame, don't force results selector or repeat peak search
The logic behind forcing the results selector (to "Calculations within
GUI") was that anything calculated in the GUI should appear under that
selection only.

Unfortunately, when loading a stream without images, there is nothing at
all to be seen under this selection, not even an image.  Forcing the
selector resulted in proj->cur_image = NULL and therefore a crash when
trying to index based on that.

Similarly, the call to update_peaks expected image data, but there is
none.  In reality, we have a perfectly good image structure with peaks
from the stream already.

This allows a non-image stream to be "re-indexed" based on the peaks.

Fixes: https://gitlab.desy.de/thomas.white/crystfel/-/issues/33