Set panel for bad region when there is only one panel
When there is only one panel, we can safely assume that fs/ss bad
regions refer to it without it being stated in the geometry file.
This allows some older Pilatus geometry files to be used without
complaints.

A further improvement would be to determine the panel when all the data
comes from one "slab" so the panel is still ambiguous.  However, on
further thought it seems like a lot of work to do this reliably in all
cases.  I'm not even sure that this case in particular is causing anyone
problems.

Closes: https://gitlab.desy.de/thomas.white/crystfel/-/issues/32