Rework header caching layer
This clears up multiple layering violations which were starting to get
in the way.  It enables "string" headers to be stored directly, and it
will make it much simpler to add new header types in the future.

Along the way, this changes all the floating point header stuff to use
double precision.  This is needed for EuXFEL event IDs.

Closes: https://gitlab.desy.de/thomas.white/crystfel/-/issues/34