GUI: Run all jobs with same CWD as GUI itself
This requires changing all the paths to relative ones.

Previously, the jobs ran with their CWDs set to the corresponding
"result" folder.  The trouble is, this breaks everything that uses a
relative location: image filenames, mask/satmap files referenced by the
geometry, etc.

Note that this change also means that the temporary folders
("indexamajig.XXX") are now produced in the top-level working directory.
Eventually, a GUI option should be added to allow these folders to be
put somewhere better (i.e. somewhere fast).

Closes: https://gitlab.desy.de/thomas.white/crystfel/-/issues/44