Increase threshold for integration warning
The warning about "XXX reflections couldn't be integrated" is tiresome -
it appears for practically every indexed pattern because there are
always some reflections that fall in bad regions or at panel edges.

This suppresses the warning unless the number of integration failures is
more than a quarter of the total number of reflections.