Remove detector rail direction altogether
This got missed out by accident in the conversion to DataTemplate, but
absolutely no-one noticed.  In the meantime, my views on how the
geometry files should work have changed somewhat.  I don't want to
maintain the extra complexity here when it isn't even clear that it will
eliminate the need to re-refine geometry for each camera length.

This commit just takes the rail direction stuff out of the documentation
and the geometry file parser.

Closes: https://gitlab.desy.de/thomas.white/crystfel/-/issues/50