SLURM: Copy environment from parent process
Unfortunately, several environments rely on "magic" environment
variables to make things work.  These need to be propagated.

The path to the GUI will no longer be added to the PATH.  This was
intended to help when the indexing executables (mosflm, dirax etc) were
found alongside the CrystFEL executables.  However, it's highly likely
that the path to the CrystFEL executables will be in PATH anyway.

This way, running the SLURM jobs in the same environment as the GUI
itself, also seems more compliant with the "principle of least
surprise".