Move file_exists() check to right before opening the HDF5
No other code path under image_expand_frames() actually opens the file,
because HDF5s are currently the only format supporting multiple frames.
The aim of the existence check is to give a better error message
(compared to the HDF5 backtrace splurge), but it doesn't need to be done
so early.  In fact, checking so early had the side-effect of breaking
indexamajig --wait-for-file.

Fixes: https://gitlab.desy.de/thomas.white/crystfel/-/issues/65