GUI: Better handling of error-filled indexamajig logs
If there are lots of error messages (e.g. a missing mask file), looking
at the last 4k of the log file is not always enough.  Now it will go
back to 16k, in steps of 4k.  It will stop early if it finds the start
of the file, and it will complain if it still doesn't find what it's
looking for.