Zero-weight negative intensity peaks in prediction refinement
Strongly negative peaks can occur if there are unmasked bad pixels.
These cause the refinement to diverge badly.  This change stabilises the
refinement such that it works even in these cases.  But of course, the
real solution is to mask out those pixels.

Negative intensity peaks are unlikely to show up "for real" in the peak
search results (because we look for strong peaks!).  However, if the
intensity is negative, we aren't even sure that the peak exists at all
and have no reason to assume it should be close to the Bragg condition.