Restructure image data array creation
This adds a central procedure (image_create_dp_bad_sat) to create all
the arrays.  Then it's up to the loading procedure to put the values
into the arrays.

This also makes the loading procedures responsible for marking NaN/inf
pixels in the bad pixel map.  This avoids an additional sweep through
the image data, and makes it possible to skip the NaN/inf check
altogether if the image data comes, as it often does, in format which
can't represent NaN/inf anyway.

Finally, it removes quite a lot of duplicated code.