Fix iterate() for prediction refinement
I was confused when I wrote acd0de4e4a21, and completely broke the
maths.  To fix it, I copied the guts of iterate() back from the old version and
re-created the intermediate steps - switching x/y to fs/ss, calculating
fs/ss gradients together and the hooks for Minv.