Julia: fix Image/PeakList memory semantics
The previous commit 3bb46f49 was a mistake.  We can't just free the old
peak list, because there might still be a PeakList object around that
refers to it.

This new version keeps a record of the PeakList object itself, setting
in_image=false if a new list is added, so that it can be freed by the
GC.

I made an attempt at making this thread-safe, but I think it's
impossible to get it completely correct.  One should simply avoid
setting any Image fields in the same structure from different threads.