Julia: Image/PeakList memory semantics (again)
Previous commit a6462e1f0 was still not right.  If the image gets freed,
the PeakList can be left with a dangling reference.  If the peak list
gets replaced at the C level (e.g. by running a new peak search), Julia
will have no way of knowing about it.

Instead of having the PeakList know if it's associated with an image,
it's better to have the Image know (at the C level) if it's responsible
for freeing the ImageFeatureList.  As soon as the ImageFeatureList is
exposed to the Julia GC via a PeakList object, it becomes Julia's
responsibility to free it.  The Julia Image structure contains a
reference to the Julia PeakList, to prevent this from happening until
either the image is freed or a new PeakList is substituted.  The two
references (Julia image.peaklist and C image->features) have to be kept
in sync, and we check image->features every time the peaklist is
requested.