Crystals shouldn't own RefLists (part 1)
This is a terrible bit of API.  A Crystal contains both the parameters
for a calculation (e.g. prediction) and the results.  Just look at
post-refinement.c for an example of the mess this makes when trying to
do calculations.

This commit removes the reflection list from the Crystal structure.
Future commits in this series will fix the resulting build carnage.

This also gets rid of vestigial field pr_dud, and adds initialisers for
all (remaining) fields.