Sort-of-workaround for argp '--help' crash on MacOS
Indexamajig --help causes a segfault on MacOS due to a bug in
argp-standalone.  That package is unlikely to be updated any time soon.
Eventually, Homebrew might switch to one of the non-crashy forks, but
probably not soon.  We aren't the only ones affected - see
https://www.openwall.com/lists/musl/2021/02/10/1

This commit adds a warning message to state that we already know about
it, and it's not "our" fault.  It also removes the corresponding test,
since we know it fails.