asdf: Force cell to be right-handed after refinement
In some rare cases, the cell changes from right to left handed during
the vector refinement stage of asdf, while ending up at an apparently
correct indexing solution.  This commit just flips the axis round when
necessary, in the same way that asdf already does in an earlier stage.

I haven't done a proper survey, but I can say that the image I used to
isolate and test this issue had very low resolution while still being
what we would call "indexable".  The left-handed cells appeared for six
out of the 12173 hits in that run, or slightly below 0.05%.  So, it's
rare but not totally impossible.