data_template_write_to_fh: Don't try to write coffset at top level
Previously, there was a bad bug here.  We would still try to write a
top-level coffset even if panels have moved such that the coffset is no
longer a single value.

One possibility is to unset panel->cnz_offset_default when moving a
panel.  This works, but we have to do it on every translation, or accept
a floating-point comparison with zero and associated non-determinism.

It's very likely that if we move one panel, we'll be moving all panels,
even if only by a small amount.  So, a more stable way seems to just
write separate coffset values all the time.  The coffset values aren't
really meant to be used at the top level anyway (that's what clen is
for!), so this fits pretty well.