compare_reindexed_cell_parameters: Don't automatically select identity matrix
The previous behaviour gave biased cell distributions in some
situations, because it preferred not to overrule the indexing algorithm.
Some indexing algorithms (xgandalf) always return a certain cell
variant, e.g. gamma > 90.  To make the histograms interpretable, we have
to randomise our choice of reduced cell in all cases.

However, we need to be careful that the cells we randomise between are
really equivalent.  The previous behaviour here was to look only at the
axis lengths.  We must look at the angles as well.

But that's not the end of the story.  We would have to choose some
weighting factor between axis lengths and angles when deciding if a new
cell is "better".  That would be hard.  Instead, compare_rcp now
compares the G6 vectors directly, which consider lengths and angles on
an even footing.

Fixes: https://gitlab.desy.de/thomas.white/crystfel/-/issues/95