Reduce minimum number of pairs for prediction refinement
This was previously being used as a kind of check for correctness of the
indexing solution.  However, the threshold only made sense for large
unit cells.  It breaks indexing for small cells, which we are
increasingly seeing in applications.

Now, the threshold is reduced to the theoretical minimum below which the
refinement is strictly underdetermined.  We (currently!) refine 9
parameters and have three data points per pair (fs,ss,excitation error).