Revert "integrate_rings_once: Handle case where reflection is negative by < 1 photon"
This special case is not necessary, because the same situation is
already handled by the within-one-photon-of-zero check just before.

The value assigned to sig2_poisson in this special case is wrong anyway:
it should be aduph*aduph.  The wrong value leads to a jump in Poisson
sigmas at low intensities.  However, the overall sigma is usually
dominated by the background variance for weak reflections, so it doesn't
make a big difference overall.

This reverts commit d4873dff09b3cf314fce0019d76e3a4ca64d7221.