Catch NaNs in ggpm model
Previously, these NaNs would result in very many reflections being
predicted with meaningless partiality when the profile radius was zero.
This would take a long time (~30 seconds) and hold up the processing of
other patterns.  A zero radius is sometimes returned by the auto profile
radius calculation, which is a separate problem.

Now, the model simply doesn't predict any reflections and returns very
quickly.

It feels strange to not have any predictions in this case, but the model
is actually giving the expected answer, because the overlap integral is
always zero when the profile radius is zero.   Therefore there can be no
intensity in the peak and hence no prediction.

Fixes: https://gitlab.desy.de/thomas.white/crystfel/-/issues/109