peakfinder8: Disable max-res cutoff by default
This changes the behaviour of peakfinder8 such that --max-res=0 disables
the high-resolution cutoff completely.  It then sets this value as the
default for indexamajig and in the GUI.

When peakfinder8 was first created, 1200 pixels was a generous cutoff.
These days, we have many detectors that are much bigger and the default
value for --max-res seems to create confusion.