indexamajig: Avoid duplicate indexer info
A consequence of fork-exec is that each worker process prints the
indexer info at startup.  We often have >100 processes, so that's a lot
of spam.  With this commit, the indexer info is printed by a separate
routine which has to be called explicitly by the top level.  Indexamajig
only does this for worker 0.

If worker 0 gets re-spawned, we'll see the message again.  That seems to
be an acceptable trade-off.

The indexer info now won't get printed by the GUI, unless we go back and
add the call to print_indexing_info.