.gitlab-ci: deploy via copy and symlink switcheroo
During analysis on Maxwell, for one of the jobs we saw error messages
indicating that "libmsgpack.so" and some xgandalf precomputed files
weren't found. The cause must have been a partial update that was
being made by the deployment script.

To fix it, we now copy and rename, instead of renaming and then
copying.