Use axis/angle formulation during prediction refinement
Previously, the prediction refinement directly used the cartesian
components of the unit cell basis vectors.  This makes it very hard to
impose Bravais restrictions, e.g. constraining all the angles to be 90°
or certain axis lengths to be the same.

This commit changes the parameterisation to instead use three axis
lengths, three angles and three overall rotation angles.  The total
number of degrees of freedom is the same, and there's little to no
effect on the results.