Observe cell restrictions during prediction refinement
This splits the prediction refinement into two stages.  In the first
step, we pretend the cell is triclinic and refine exactly as before
(except that we use the axis/angle formulation instead of the cartesian
components).  The lattice conditions are then imposed on the cell, e.g.
setting angles to 90°, and another round of refinement made while
maintaining those constraints.