Fix semantics of data_template_get_[2d_]detgeom_if_possible()
The routine tried to do two things:
1. Create a detgeom for a flat detector only
2. Create a detgeom ignoring the camera length

Now, the one routine that is interested in the flatness of the detector
(data_template_get_clen_if_possible) checks itself, and
data_template_get_detgeom_if_possible (note removal of '2d_') has an
extra parameter to say whether it's OK if we don't know the overall
camera length.

This might have made more sense when panels could have individual clens
(as well as coffsets).  But now that clen is a top-level parameter, and
coffset is always a literal number, things make more sense.

Fixes: https://gitlab.desy.de/thomas.white/crystfel/-/issues/127