indexamajig: Avoid shared queue when using ASAP::O
This splits the queue access into a separate routine, which waits on the
semaphore and then retrieves the filename, frame ID and serial number.
When being fed data via ASAP::O, we have an alternative (better) source
for all of these, so there's no need for the queue and its associated
lock contention.