Meson: Check for 'ccp4c' as well as 'libccp4c'
Between 6.5.1 and 8.0.0, libccp4 changed the pkg-config name from
'libccp4c' to 'ccp4c'.  The Meson wrap has not (yet) been updated to
reflect this.

Searching for both versions allows us to use either version (6.5.1 or
8.0.0), as well as the Meson wrapped version (e.g. from a previous
CrystFEL installation).

This increases the minimum Meson version we require from 0.55.0 to 0.60.0.
Unfortunately, this looks to be the final nail in the coffin for us
supporting Ubuntu Xenial (16.04).  This distribution will be fully
end-of-life (no security updates) in April, so I don't see this as a big
loss.  Enterprising users working on ancient distributions can still
make things work by reverting this change.