CI: Build the current branch on MacOS
The previous version tested the Homebrew formula (now migrated to a
private tap), which always used the master branch.

Now, there are two CI jobs.  One builds "manually", with dependencies
from Homebrew but nothing more, on the current branch.  The other one
runs only for the master branch and tests the build using the formula
from the tap.